chore: refatored some code & updated README (#65)
* refactor: Update workflow files and rename Rust tests workflow (#63)

* refactor: Update workflow files and rename Rust tests workflow

* Logger (#64)

* chore: Update slog dependency to version 2.7.0 with max_level_trace feature

* chore: renamed logger to logging

* refactor: Update logger imports to use the new logging module

* chore: Update dependencies for logging and slog

+++ tokenscout/src/cli/mod.rs
@@ -1,12 +1,12 @@
 use slog::*;
 
 use clap::{Arg, ArgAction, Command};
-use logger::parse_verbosity;
+use logging::parse_verbosity;
 use slog::{crit, debug, error, info, trace, warn};
 use tokenscout_version::VERSION;
 
 pub struct TokenscoutCli {
-    pub verbosity: u64,
+    pub verbosity: u8,
     pub chains: Vec<String>,
     pub telegram_channels: Vec<String>,
     pub custom_endpoint: Option<Vec<String>>,
